Finetuned speed and distance travelled

+++ lib/roadfighter/Entity.h
@@ -30,7 +30,7 @@ namespace roadfighter {
 
         virtual int getMax_speed() const {return 0;};
 
-        virtual int getDistance() const {return 0;};
+        virtual double getDistance() const {return 0;};
 
         virtual void update_distance(double distance) {};
     };
